added item transfer helper to reduce code duplication
